requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
kombu==5.5.4
numpy==2.3.3
openai==1.108.0
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.11.0
